@@ -6,5 +6,15 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._automanage_client_async import AutomanageClient
+from ._automanage_client import AutomanageClient
+
+try:
+    from ._patch import __all__ as _patch_all
+    from ._patch import *  # type: ignore # pylint: disable=unused-wildcard-import
+except ImportError:
+    _patch_all = []
+from ._patch import patch_sdk as _patch_sdk
 __all__ = ['AutomanageClient']
+__all__.extend([p for p in _patch_all if p not in __all__])
+
+_patch_sdk()
