@@ -7,58 +7,58 @@
 # --------------------------------------------------------------------------
 
 from copy import deepcopy
-from typing import Any, TYPE_CHECKING
+from typing import Any, Awaitable, TYPE_CHECKING
 
 from msrest import Deserializer, Serializer
 
-from azure.core.rest import HttpRequest, HttpResponse
-from azure.mgmt.core import ARMPipelineClient
+from azure.core.rest import AsyncHttpResponse, HttpRequest
+from azure.mgmt.core import AsyncARMPipelineClient
 
-from . import models
+from .. import models
 from ._configuration import AutomanageClientConfiguration
 from .operations import BestPracticesOperations, BestPracticesVersionsOperations, ConfigurationProfileAssignmentsOperations, ConfigurationProfileHCIAssignmentsOperations, ConfigurationProfileHCRPAssignmentsOperations, ConfigurationProfilesOperations, ConfigurationProfilesVersionsOperations, HCIReportsOperations, HCRPReportsOperations, Operations, ReportsOperations, ServicePrincipalsOperations
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from azure.core.credentials import TokenCredential
+    from azure.core.credentials_async import AsyncTokenCredential
 
 class AutomanageClient:    # pylint: disable=too-many-instance-attributes
     """Automanage Client.
 
     :ivar best_practices: BestPracticesOperations operations
-    :vartype best_practices: azure.mgmt.automanage.operations.BestPracticesOperations
+    :vartype best_practices: azure.mgmt.automanage.aio.operations.BestPracticesOperations
     :ivar best_practices_versions: BestPracticesVersionsOperations operations
     :vartype best_practices_versions:
-     azure.mgmt.automanage.operations.BestPracticesVersionsOperations
+     azure.mgmt.automanage.aio.operations.BestPracticesVersionsOperations
     :ivar configuration_profiles: ConfigurationProfilesOperations operations
     :vartype configuration_profiles:
-     azure.mgmt.automanage.operations.ConfigurationProfilesOperations
+     azure.mgmt.automanage.aio.operations.ConfigurationProfilesOperations
     :ivar configuration_profiles_versions: ConfigurationProfilesVersionsOperations operations
     :vartype configuration_profiles_versions:
-     azure.mgmt.automanage.operations.ConfigurationProfilesVersionsOperations
+     azure.mgmt.automanage.aio.operations.ConfigurationProfilesVersionsOperations
     :ivar configuration_profile_assignments: ConfigurationProfileAssignmentsOperations operations
     :vartype configuration_profile_assignments:
-     azure.mgmt.automanage.operations.ConfigurationProfileAssignmentsOperations
+     azure.mgmt.automanage.aio.operations.ConfigurationProfileAssignmentsOperations
     :ivar operations: Operations operations
-    :vartype operations: azure.mgmt.automanage.operations.Operations
+    :vartype operations: azure.mgmt.automanage.aio.operations.Operations
     :ivar reports: ReportsOperations operations
-    :vartype reports: azure.mgmt.automanage.operations.ReportsOperations
+    :vartype reports: azure.mgmt.automanage.aio.operations.ReportsOperations
     :ivar service_principals: ServicePrincipalsOperations operations
-    :vartype service_principals: azure.mgmt.automanage.operations.ServicePrincipalsOperations
+    :vartype service_principals: azure.mgmt.automanage.aio.operations.ServicePrincipalsOperations
     :ivar configuration_profile_hcrp_assignments: ConfigurationProfileHCRPAssignmentsOperations
      operations
     :vartype configuration_profile_hcrp_assignments:
-     azure.mgmt.automanage.operations.ConfigurationProfileHCRPAssignmentsOperations
+     azure.mgmt.automanage.aio.operations.ConfigurationProfileHCRPAssignmentsOperations
     :ivar hcrp_reports: HCRPReportsOperations operations
-    :vartype hcrp_reports: azure.mgmt.automanage.operations.HCRPReportsOperations
+    :vartype hcrp_reports: azure.mgmt.automanage.aio.operations.HCRPReportsOperations
     :ivar configuration_profile_hci_assignments: ConfigurationProfileHCIAssignmentsOperations
      operations
     :vartype configuration_profile_hci_assignments:
-     azure.mgmt.automanage.operations.ConfigurationProfileHCIAssignmentsOperations
+     azure.mgmt.automanage.aio.operations.ConfigurationProfileHCIAssignmentsOperations
     :ivar hci_reports: HCIReportsOperations operations
-    :vartype hci_reports: azure.mgmt.automanage.operations.HCIReportsOperations
+    :vartype hci_reports: azure.mgmt.automanage.aio.operations.HCIReportsOperations
     :param credential: Credential needed for the client to connect to Azure.
-    :type credential: ~azure.core.credentials.TokenCredential
+    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: The ID of the target subscription.
     :type subscription_id: str
     :param base_url: Service URL. Default value is "https://management.azure.com".
@@ -70,13 +70,13 @@ class AutomanageClient:    # pylint: disable=too-many-instance-attributes
 
     def __init__(
         self,
-        credential: "TokenCredential",
+        credential: "AsyncTokenCredential",
         subscription_id: str,
         base_url: str = "https://management.azure.com",
         **kwargs: Any
     ) -> None:
         self._config = AutomanageClientConfiguration(credential=credential, subscription_id=subscription_id, **kwargs)
-        self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+        self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -124,14 +124,14 @@ class AutomanageClient:    # pylint: disable=too-many-instance-attributes
         self,
         request: HttpRequest,
         **kwargs: Any
-    ) -> HttpResponse:
+    ) -> Awaitable[AsyncHttpResponse]:
         """Runs the network request through the client's chained policies.
 
         >>> from azure.core.rest import HttpRequest
         >>> request = HttpRequest("GET", "https://www.example.org/")
         <HttpRequest [GET], url: 'https://www.example.org/'>
-        >>> response = client._send_request(request)
-        <HttpResponse: 200 OK>
+        >>> response = await client._send_request(request)
+        <AsyncHttpResponse: 200 OK>
 
         For more information on this code flow, see https://aka.ms/azsdk/python/protocol/quickstart
 
@@ -139,22 +139,19 @@ class AutomanageClient:    # pylint: disable=too-many-instance-attributes
         :type request: ~azure.core.rest.HttpRequest
         :keyword bool stream: Whether the response payload will be streamed. Defaults to False.
         :return: The response of your network call. Does not do error handling on your response.
-        :rtype: ~azure.core.rest.HttpResponse
+        :rtype: ~azure.core.rest.AsyncHttpResponse
         """
 
         request_copy = deepcopy(request)
         request_copy.url = self._client.format_url(request_copy.url)
         return self._client.send_request(request_copy, **kwargs)
 
-    def close(self):
-        # type: () -> None
-        self._client.close()
+    async def close(self) -> None:
+        await self._client.close()
 
-    def __enter__(self):
-        # type: () -> AutomanageClient
-        self._client.__enter__()
+    async def __aenter__(self) -> "AutomanageClient":
+        await self._client.__aenter__()
         return self
 
-    def __exit__(self, *exc_details):
-        # type: (Any) -> None
-        self._client.__exit__(*exc_details)
+    async def __aexit__(self, *exc_details) -> None:
+        await self._client.__aexit__(*exc_details)
